+++ src/daemon/session-target-evidence.ts
@@ -23,8 +23,12 @@ import {
   siblingOrdinal,
 } from '../replay/target-identity-node.ts';
 import {
+  buildAncestryChain,
+  buildIndexMap,
+  filterIdentitySet,
+} from '../replay/target-evidence-tree.ts';
+import {
   classifyTargetBindingMatch,
-  matchesAncestryPrefix,
   matchesLocalIdentity,
   serializeTargetAnnotationV1,
   utf8ByteLength,
@@ -118,48 +122,8 @@ export function computeTargetEvidence(
 // primitives, computed once per candidate node against the record-time tree.
 // ---------------------------------------------------------------------------
 
-export function buildIndexMap(nodes: readonly SnapshotNode[]): Map<number, SnapshotNode> {
-  const map = new Map<number, SnapshotNode>();
-  for (const node of nodes) map.set(node.index, node);
-  return map;
-}
-
 /** The one identity reader (normalized AND field-capped, on every path): shared with dispatch's post-resolution guard. */
 export const boundedLocalIdentity = readNodeLocalIdentity;
-
-export type AncestryWalk = {
-  chain: TargetAncestryEntry[];
-  /**
-   * Decision 3 capture anomaly: a `parentIndex` that resolves to no node, or
-   * a parent cycle. A broken walk fails the annotation closed to
-   * `unverifiable` — the structural signals cannot be trusted.
-   */
-  broken: boolean;
-};
-
-/** Decision 3 "Ancestry": nearest K ancestors, leaf→root, {role,label?}. */
-export function buildAncestryChain(
-  node: SnapshotNode,
-  byIndex: Map<number, SnapshotNode>,
-  limit: number,
-): AncestryWalk {
-  const chain: TargetAncestryEntry[] = [];
-  const visited = new Set<number>([node.index]);
-  let current = node;
-  while (chain.length < limit) {
-    if (typeof current.parentIndex !== 'number') return { chain, broken: false };
-    const parent = byIndex.get(current.parentIndex);
-    if (!parent || visited.has(parent.index)) return { chain, broken: true };
-    visited.add(parent.index);
-    const identity = boundedLocalIdentity(parent);
-    chain.push({
-      role: identity.role,
-      ...(identity.label !== undefined ? { label: identity.label } : {}),
-    });
-    current = parent;
-  }
-  return { chain, broken: false };
-}
 
 /**
  * Decision 3 record-time write step 3: the winner's zero-based index among
@@ -218,28 +182,6 @@ type DisambiguationDomain = {
   orderedRegion: SnapshotNode[];
   viewportOrder: number;
 };
-
-/**
- * Decision 3 record-time write step 2 / replay-time verification's identity
- * set I: candidates sharing the recorded local identity with a matching
- * leaf-anchored ancestry prefix. Shared by the writer's self-check (over the
- * full record-time tree) and replay-time verification (over the recorded
- * selector/ref's matched-node domain) — "record and replay compute this
- * ordinal identically by definition" applies to this filter too.
- */
-export function filterIdentitySet(
-  candidates: readonly SnapshotNode[],
-  byIndex: Map<number, SnapshotNode>,
-  identity: LocalIdentity,
-  ancestry: readonly TargetAncestryEntry[],
-): SnapshotNode[] {
-  return candidates.filter((candidate) => {
-    if (!matchesLocalIdentity(boundedLocalIdentity(candidate), identity)) return false;
-    const observed = buildAncestryChain(candidate, byIndex, Math.max(ancestry.length, 1));
-    // A candidate with a broken parent walk cannot prove the prefix.
-    return !observed.broken && matchesAncestryPrefix(observed.chain, ancestry);
-  });
-}
 
 /**
  * ADR 0012 decision 3 amendment (#1269): an id is identity only when it
